refactor(datawell): remove .js exts on imports
